+++ src/App.css
@@ -52,3 +52,7 @@
     width: 50vw;
   }
 }
+.App {
+  font-family: sans-serif;
+  text-align: center;
+}
